+++ CHANGES.rst
@@ -9,6 +9,8 @@ Unreleased
 
     -   Request and response mixins have all been merged into the
         ``Request`` and ``Response`` classes.
+    -   The ``has_key`` method on some datastructures; use
+        ``key in data`` instead.
 
 -   Default values passed to ``Headers`` are validated the same way
     values added later are. :issue:`1608`